+++ src/main/java/com/github/theredbrain/foodoverhaul/FoodOverhaul.java
@@ -1,21 +1,12 @@
 package com.github.theredbrain.foodoverhaul;
 
+import com.github.theredbrain.foodoverhaul.component.type.OverhauledFoodComponent;
 import com.github.theredbrain.foodoverhaul.config.ServerConfig;
-import com.github.theredbrain.foodoverhaul.config.ServerConfigWrapper;
-import com.github.theredbrain.foodoverhaul.registry.EventsRegistry;
-import com.github.theredbrain.foodoverhaul.registry.StatusEffectsRegistry;
-import com.google.gson.Gson;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
-import me.shedaniel.autoconfig.serializer.PartitioningSerializer;
+import com.github.theredbrain.foodoverhaul.registry.ItemComponentRegistry;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.component.ComponentType;
 import net.minecraft.entity.attribute.EntityAttribute;
-import net.minecraft.network.RegistryByteBuf;
-import net.minecraft.network.codec.PacketCodec;
-import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
@@ -24,44 +15,16 @@ import org.slf4j.LoggerFactory;
 public class FoodOverhaul implements ModInitializer {
 	public static final String MOD_ID = "foodoverhaul";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-	public static ServerConfig serverConfig;
+	public static ServerConfig SERVER_CONFIG = ConfigApiJava.registerAndLoadConfig(ServerConfig::new);
 
+	public static ComponentType<OverhauledFoodComponent> OVERHAULED_FOOD_COMPONENT_TYPE;
 	public static RegistryEntry<EntityAttribute> MAX_FOOD_EFFECTS;
 
 	@Override
 	public void onInitialize() {
 		LOGGER.info("Enjoy your overhauled food!");
 
-		// Config
-		AutoConfig.register(ServerConfigWrapper.class, PartitioningSerializer.wrap(JanksonConfigSerializer::new));
-		serverConfig = ((ServerConfigWrapper) AutoConfig.getConfigHolder(ServerConfigWrapper.class).getConfig()).server;
-
-		PayloadTypeRegistry.playS2C().register(ServerConfigSyncPacket.PACKET_ID, ServerConfigSyncPacket.PACKET_CODEC);
-		ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
-			ServerPlayNetworking.send(handler.player, new ServerConfigSyncPacket(serverConfig));
-		});
-
-		StatusEffectsRegistry.init();
-
-		EventsRegistry.initializeEvents();
-	}
-
-	public record ServerConfigSyncPacket(ServerConfig serverConfig) implements CustomPayload {
-		public static final CustomPayload.Id<ServerConfigSyncPacket> PACKET_ID = new CustomPayload.Id<>(identifier("server_config_sync"));
-		public static final PacketCodec<RegistryByteBuf, ServerConfigSyncPacket> PACKET_CODEC = PacketCodec.of(ServerConfigSyncPacket::write, ServerConfigSyncPacket::new);
-
-		public ServerConfigSyncPacket(RegistryByteBuf registryByteBuf) {
-			this(new Gson().fromJson(registryByteBuf.readString(), ServerConfig.class));
-		}
-
-		private void write(RegistryByteBuf registryByteBuf) {
-			registryByteBuf.writeString(new Gson().toJson(serverConfig));
-		}
-
-		@Override
-		public CustomPayload.Id<? extends CustomPayload> getId() {
-			return PACKET_ID;
-		}
+		ItemComponentRegistry.init();
 	}
 
 	public static Identifier identifier(String path) {