+++ src/main/java/com/github/theredbrain/foodoverhaul/mixin/item/ItemMixin.java
@@ -1,8 +1,8 @@
 package com.github.theredbrain.foodoverhaul.mixin.item;
 
+import com.github.theredbrain.foodoverhaul.FoodOverhaul;
+import com.github.theredbrain.foodoverhaul.component.type.OverhauledFoodComponent;
 import com.github.theredbrain.foodoverhaul.entity.player.DuckPlayerEntityMixin;
-import net.minecraft.component.DataComponentTypes;
-import net.minecraft.component.type.FoodComponent;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -20,8 +20,8 @@ public abstract class ItemMixin {
 	@Inject(method = "use", at = @At("HEAD"), cancellable = true)
 	public void foodoverhaul$use(World world, PlayerEntity user, Hand hand, CallbackInfoReturnable<TypedActionResult<ItemStack>> cir) {
 		ItemStack itemStack = user.getStackInHand(hand);
-		FoodComponent foodComponent = itemStack.get(DataComponentTypes.FOOD);
-		if (foodComponent != null) {
+		OverhauledFoodComponent overhauledFoodComponent = itemStack.get(FoodOverhaul.OVERHAULED_FOOD_COMPONENT_TYPE);
+		if (overhauledFoodComponent != null) {
 			if (((DuckPlayerEntityMixin) user).foodoverhaul$canConsumeItem(itemStack)) {
 				user.setCurrentHand(hand);
 				cir.setReturnValue(TypedActionResult.consume(itemStack));