+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.2.0
+
+- split food effects into standalone mod
+- removed dependency on cloth config
+- added dependency on fzzy config
+- added custom item component
+
 # 2.1.0
 
 - update to 1.21.1